+++ app/src/main/java/com/example/gear_kotlin/ui/MainActivity.kt
@@ -1,10 +1,10 @@
-package com.example.gear_kotlin
+package com.example.gear_kotlin.ui
 
-import android.content.Intent
 import android.os.Bundle
 import android.view.Menu
 import android.view.MenuItem
 import androidx.appcompat.app.AppCompatActivity
+import com.example.gear_kotlin.R
 
 import kotlinx.android.synthetic.main.activity_main.*
