+++ app/src/main/res/layout/app_fragment_first.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AppFirstFragment">
+    tools:context=".ui.AppFirstFragment">
 
     <TextView
         android:id="@+id/textview_first"
